Changed vector output to show only ints.
git-svn-id: svn://svn.code.sf.net/p/toscanaj/code/trunk@544 267469b6-68fb-4578-8a4b-5844dc4013a7

+++ source/net/sourceforge/toscanaj/view/diagram/NodeView.java
@@ -117,7 +117,7 @@ public class NodeView extends CanvasItem {
                 vector = "(";
                 for (int i = 0; i < ndimVec.length; i++) {
                     double v = ndimVec[i];
-                    vector += v;
+                    vector += (int)v;
                     if(i != ndimVec.length-1) {
                         vector += ",";
                     }
